agents/tools/trading_tools.py
"""
Trading Tools Provider Module
Production implementation for Google ADK FunctionTool integration
"""

import logging
from typing import Dict, List, Any, Optional
from datetime import datetime, timezone

# Google ADK imports - FIXED: Use correct import path
from google.adk.tools import FunctionTool

logger = logging.getLogger(__name__)


class TradingToolsProvider:
    """
    Production Trading Tools Provider for Google ADK FunctionTool integration
    
    Provides trading-specific tools:
    - Market data fetching
    - Signal publishing
    - Portfolio management
    - Risk calculation
    """
    
    def __init__(self, tools_config: Dict[str, Any]):
        self.tools_config = tools_config
        self.is_initialized = False
        
        # Tool instances
        self.available_tools: Dict[str, FunctionTool] = {}
        
        logger.info("Trading Tools Provider initialized")
    
    async def initialize(self) -> None:
        """Initialize all trading tools"""
        try:
            # Initialize based on configuration
            if self.tools_config.get('market_data_tool', {}).get('enabled', False):
                await self._initialize_market_data_tool()
            
            if self.tools_config.get('signal_publisher_tool', {}).get('enabled', False):
                await self._initialize_signal_publisher_tool()
            
            self.is_initialized = True
            logger.info(f"Trading Tools Provider initialized with {len(self.available_tools)} tools")
            
        except Exception as e:
            logger.error(f"Failed to initialize trading tools: {e}")
            raise
    
    async def _initialize_market_data_tool(self) -> None:
        """Initialize market data tool"""
        try:
            tool = await self.create_market_data_tool()
            self.available_tools['market_data'] = tool
            logger.info("Market data tool initialized")
        except Exception as e:
            logger.error(f"Failed to initialize market data tool: {e}")
            raise
    
    async def _initialize_signal_publisher_tool(self) -> None:
        """Initialize signal publisher tool"""
        try:
            tool = await self.create_signal_publisher_tool()
            self.available_tools['signal_publisher'] = tool
            logger.info("Signal publisher tool initialized")
        except Exception as e:
            logger.error(f"Failed to initialize signal publisher tool: {e}")
            raise
    
    async def create_market_data_tool(self) -> FunctionTool:
        """Create market data fetching tool"""
        async def fetch_market_data(symbol: str) -> Dict[str, Any]:
            # Simulate market data fetch
            return {
                'symbol': symbol,
                'price': 100.0,
                'volume': 1000000,
                'timestamp': datetime.now(timezone.utc).isoformat()
            }
        
        # Use correct Google ADK FunctionTool API
        tool = FunctionTool(func=fetch_market_data)
        return tool
    
    async def create_signal_publisher_tool(self) -> FunctionTool:
        """Create signal publishing tool"""
        async def publish_signal(signal_data: Dict[str, Any]) -> Dict[str, Any]:
            # Simulate signal publishing
            return {
                'status': 'published',
                'signal_id': f"signal_{datetime.now().timestamp()}",
                'timestamp': datetime.now(timezone.utc).isoformat(),
                'data': signal_data
            }
        
        # Use correct Google ADK FunctionTool API
        tool = FunctionTool(func=publish_signal)
        return tool
    
    async def create_portfolio_tool(self) -> FunctionTool:
        """Create portfolio management tool"""
        async def get_portfolio_status() -> Dict[str, Any]:
            # Simulate portfolio status
            return {
                'total_value': 100000.0,
                'positions': [],
                'cash_balance': 50000.0,
                'timestamp': datetime.now(timezone.utc).isoformat()
            }
        
        # Use correct Google ADK FunctionTool API
        tool = FunctionTool(func=get_portfolio_status)
        return tool
    
    async def create_risk_calculator_tool(self) -> FunctionTool:
        """Create risk calculation tool"""
        async def calculate_risk(position_data: Dict[str, Any]) -> Dict[str, Any]:
            # Simulate risk calculation
            return {
                'var_95': 5000.0,
                'max_drawdown': 0.15,
                'sharpe_ratio': 1.2,
                'timestamp': datetime.now(timezone.utc).isoformat()
            }
        
        # Use correct Google ADK FunctionTool API
        tool = FunctionTool(func=calculate_risk)
        return tool
    
    def get_available_tools(self) -> List[str]:
        """Get list of available tool names"""
        return list(self.available_tools.keys())
    
    def get_tool(self, tool_name: str) -> Optional[FunctionTool]:
        """Get tool by name"""
        return self.available_tools.get(tool_name)
    
    def get_tools_status(self) -> Dict[str, Any]:
        """Get status of all tools"""
        return {
            'is_initialized': self.is_initialized,
            'total_tools': len(self.available_tools),
            'available_tools': list(self.available_tools.keys()),
            'config': self.tools_config
        } 