requirements.txt
# FastAPI Core Dependencies
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.12

# Authentication & Security
python-jose[cryptography]==3.3.0
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
pyotp==2.9.0
qrcode[pil]==7.4.2
cryptography==42.0.8

# Database & ORM
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
asyncpg==0.29.0
alembic==1.14.0

# Redis & Caching
redis==5.2.1
aioredis==2.0.1

# Environment & Configuration
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0

# HTTP Client
httpx==0.28.1
aiohttp==3.11.11

# Validation & Utilities
email-validator==2.2.0

# Rate Limiting & Circuit Breaker
slowapi==0.1.9
asyncio-throttle==1.0.2

# Vector Database & Search
qdrant-client==1.11.3
sentence-transformers==3.2.1

# Message Queue & Data Processing
pulsar-client==3.4.0
elasticsearch==8.15.1

# Data Processing & ML
pandas==2.1.4
numpy>=1.26.0
spacy==3.7.2
scikit-learn==1.3.2
transformers>=4.41.0
torch>=2.6.0

# Google ADK & Multi-Agent Framework
google-adk==1.5.0
litellm==1.73.2
google-genai==1.23.0
google-auth==2.40.3
google-api-python-client==2.174.0

# WebSocket & Real-time
websockets==12.0

# Data Ingestion & APIs
feedparser==6.0.10
tweepy==4.14.0
praw==7.7.1

# Data Validation
great-expectations==0.18.19

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-cov==6.0.0
pytest-order==1.3.0

# Monitoring & Logging
structlog==24.4.0
prometheus_client==0.22.1
grafana-api==1.0.3
psutil==7.0.0
docker==7.1.0

# OpenTelemetry Observability & Tracing
opentelemetry-api>=1.31.0
opentelemetry-sdk>=1.31.0
opentelemetry-exporter-otlp>=1.31.0
opentelemetry-exporter-jaeger>=1.21.0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-instrumentation-sqlalchemy>=0.48b0
opentelemetry-instrumentation-redis>=0.48b0
opentelemetry-instrumentation-httpx>=0.48b0
opentelemetry-propagator-b3>=1.31.0

# API Documentation
markdown==3.7

# MLOps Dependencies - Task 7
mlflow==2.14.1
evidently==0.4.12
apache-airflow==2.9.3
locust==2.29.1
scipy==1.12.0

# Immutable Audit Logging & Compliance Reporting
minio==7.2.0
reportlab==4.0.0
python-jose==3.3.0
hmac==0.0.1 